close NNTP after 55 seconds
and reopen the next time the connection is used. This fixes the issue
when attempting to re-use a connection that has been idle for a while.
Could probably get an event when the connection is closed by the server,
but that's in the NNTP library. Also, it's nicer to close a connection
that is not being used actively.

55 seconds is arbitrary: slightly under a minute, and just a guess.