show count "new" or last article "ago", per group
This depends on visiting the group to fetch the overviews, which sets
the last article number and date in localStorage.

Didn't find a simple event or hook to run client-side that wouldn't also
cause the group list to re-render, so component manually injects a
script element with the one-liner. Backticks for static strings avoids
the auto-"-to-&quot; conversion.

Move number formatting to format.ts. Formatting now relies on
Accept-Language as well. Tried to make formatter from method in
MyCookies, but methods do not rehydrate on client-side; only values. So
use static method.

Add timestamp to log entries.