paginate groups' article list
The back-end will attempt multiple NNTP calls to fill the requested
rows. Currently, this must complete before the race, or everything is
discarded. Would be better to go ahead with whatever is complete before
the timeout. This would likely require moving the fill loop to
WrappedBack, and some fiddling with the Promises.

Also use the (to-be-deprecated) Head component from $fresh/runtime to
set the page titles.