support multiple NNTP servers
including when the same-name but actually different group is on each.

The group in the path is now of the form `group@origin` which is
definitely wordier, but is consistently shareable in an URL (whether or
not one is using a single server at a given moment).

Also make the new message count italic so it stands out a bit more.